// src/app/Interfaces/Global/TipoDocumentoIdentidad.ts
export interface TipoDocumentoIdentidad {
  id: number;
  descripcion: string;
  abreviatura: string;
}

export interface TipoDocumentoIdentidadPost {
  descripcion: string;
  abreviatura: string;
}

export interface TipoDocumentoIdentidadPut {
  id: number;
  descripcion: string;
  abreviatura: string;
}